// src/app/Models/Project.ts
/**
 *
 */
export class Project
{

    id: number;
    team_leader_id: any = '';
    name: string;
    technologies: string;
    user_id: any;
    url: string;
    member_email: string;
    is_active: number;
    is_closed: number;
    created_at: string;
    created_by: number;
    updated_at: string;
    updated_by: number;

}